Implement launching of access point
This is just ported sample code from
https://github.com/espressif/esp-idf/blob/master/examples/wifi/getting_started/softAP/main/softap_example_main.c

It is working, meaning the specified AP is started successfully and
accepts connections. Has to be integrated into the actual flow of the
networking component.

This commit fails cpplint because of the C-style casting. See next commit
for the fix of pre-commit configuration.