Create file structure to include breathe directives
This is a first shot at #6, but does not yet really solve the issue.
Actually the entities can be correctly parsed from the source code by
Doxygen. Breathe does pick them up, but as of now, the required rST files
for Sphinx could not be created to really substitute Doxygen's generated
HTML output.