Track number of failed connections and start access point
This code is basically working, but has a serious flaw:
There are dedicated event handlers for access point and station mode,
which are registered during the startup of the respective mode.
However, unregistering does not work and results in a
``Guru meditation`` error.

Guessing: The event handler is unregistered from the event handler itsself,
which somehow corrupts the stack and then the core tries to read an
instruction from illegal memory address.

Instead of tracking this down and fixing the actual problem, let's mitigate
this by providing a common event handler for all wifi-related events, that
is shared by access point and station mode. This makes the need for
unregistering obsolete and should be pretty fast, because the wifi-related
events are distinct between these modes.

Might even save some bytes...