Start HTTP stream
This commit includes several ugly and hackerish fixes.
Obviously, a HTTP stream requires actual access to the internet, but as of
now, map32 does not know if there is internet connectivity before it is
setting up the audio_source (http_stream in this case).

Obviously, map32 needs to know, if external peripheral devices (like
internet connection, but also SD reader, ...) are ready before using them
in its own codebase.

I.e. in the example code, there is a call to the function
``periph_wifi_wait_for_connected(wifi_handle, portMAX_DELAY);``, which
actually blocks until the WiFi is ready.

This needs a lot of investigation!